# backend/nosd/openapi.yaml
openapi: 3.0.3
info:
  title: NithronOS API
  version: 0.1.0
  description: API for nOS backend
servers:
  - url: http://127.0.0.1:9000
paths:
...
  /api/pools/{id}/snapshots:
    get:
      summary: List snapshots
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Snapshot'
    post:
      summary: Create snapshot
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                subvol: { type: string }
                name: { type: string }
              required: [subvol]
      responses:
        '200':
          description: OK
...
  /api/smb/users:
    get:
      summary: List SMB users (proxy)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StringArray'
    post:
      summary: Create SMB user (proxy)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SmbUserCreate'
      responses:
        '201':
          description: Created
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/updates/apply:
    post:
      summary: Apply updates, optionally snapshotting targets before
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                packages:
                  type: array
                  items: { type: string }
                snapshot:
                  type: boolean
                confirm:
                  type: string
              required: [confirm]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  tx_id: { type: string }
                  snapshots_count: { type: integer }
                  updates_count: { type: integer }
                required: [ok, tx_id]
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/firewall/apply:
    post:
      summary: Apply firewall rules for a mode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mode: { type: string, enum: [lan-only, vpn-only, tunnel, direct] }
                twoFactorToken: { type: string, nullable: true }
              required: [mode]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                required: [ok]
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    StringArray:
      type: array
      items: { type: string }
    SmbUserCreate:
      type: object
      properties:
        username: { type: string }
        password: { type: string, nullable: true }
      required: [username]
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Snapshot:
      type: object
      properties:
        path: { type: string }
        name: { type: string }
        readonly: { type: boolean }
...